refactor: move Range struct into dedicated file
