+++ packages_rs/nextclade/src/analyze/aa_sub.rs
@@ -4,7 +4,7 @@ use crate::io::aa::{from_aa, Aa};
 use crate::io::letter::Letter;
 use crate::io::parse_pos::parse_pos;
 use crate::make_error;
-use crate::utils::range::AaRefPosition;
+use crate::utils::position::AaRefPosition;
 use eyre::{Report, WrapErr};
 use lazy_static::lazy_static;
 use regex::Regex;
