+++ packages_rs/nextclade/src/analyze/is_sequenced.rs
@@ -1,5 +1,6 @@
 use crate::analyze::letter_ranges::{GeneAaRange, NucRange};
-use crate::utils::range::{AaRefPosition, NucRefGlobalPosition, NucRefGlobalRange};
+use crate::utils::position::{AaRefPosition, NucRefGlobalPosition};
+use crate::utils::range::NucRefGlobalRange;
 use itertools::Itertools;
 
 /// Decides whether a given position in nucleotide sequence is considered "sequenced".
