+++ packages_rs/nextclade/src/analyze/letter_ranges.rs
@@ -2,7 +2,8 @@ use crate::io::aa::Aa;
 use crate::io::letter::Letter;
 use crate::io::nuc::Nuc;
 use crate::translate::translate_genes::{CdsTranslation, Translation};
-use crate::utils::range::{AaRefPosition, NucRefGlobalPosition, PositionLike, Range};
+use crate::utils::position::{AaRefPosition, NucRefGlobalPosition, PositionLike};
+use crate::utils::range::Range;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize, schemars::JsonSchema)]
