+++ packages_rs/nextclade/src/analyze/virus_properties.rs
@@ -5,7 +5,8 @@ use crate::io::fs::read_file_to_string;
 use crate::io::json::json_parse;
 use crate::io::letter::Letter;
 use crate::io::nuc::Nuc;
-use crate::utils::range::{AaRefPosition, AaRefRange, NucRefGlobalRange};
+use crate::utils::position::AaRefPosition;
+use crate::utils::range::{AaRefRange, NucRefGlobalRange};
 use eyre::{Report, WrapErr};
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
