+++ packages_rs/nextclade/src/gene/phase.rs
@@ -1,5 +1,5 @@
 use crate::make_internal_error;
-use crate::utils::range::{NucRefLocalPosition, PositionLike};
+use crate::utils::position::{NucRefLocalPosition, PositionLike};
 use eyre::Report;
 use schemars::gen::SchemaGenerator;
 use schemars::schema::Schema;
