+++ packages_rs/nextclade/src/translate/aa_alignment_ranges.rs
@@ -1,7 +1,8 @@
 use crate::io::gene_map::GeneMap;
 use crate::translate::coord_map::local_to_codon_range_exclusive;
 use crate::translate::translate_genes::Translation;
-use crate::utils::range::{intersect, AaRefRange, NucRefGlobalRange, NucRefLocalRange, PositionLike};
+use crate::utils::position::PositionLike;
+use crate::utils::range::{intersect, AaRefRange, NucRefGlobalRange, NucRefLocalRange};
 use eyre::Report;
 use itertools::Itertools;
