+++ packages_rs/nextclade/src/translate/translate_genes.rs
@@ -16,7 +16,8 @@ use crate::translate::translate::translate;
 use crate::types::outputs::PeptideWarning;
 use crate::utils::collections::{first, last};
 use crate::utils::error::report_to_string;
-use crate::utils::range::{AaRefRange, PositionLike, Range};
+use crate::utils::position::PositionLike;
+use crate::utils::range::{AaRefRange, Range};
 use crate::{make_error, make_internal_report};
 use eyre::Report;
 use indexmap::IndexMap;
