+++ packages_rs/nextclade/src/utils/mod.rs
@@ -5,6 +5,7 @@ pub mod getenv;
 pub mod global_init;
 pub mod num;
 pub mod option;
+pub mod position;
 pub mod range;
 pub mod string;
 pub mod vec2d;
